satellite_producer/requirements.txt
kafka-python==2.0.2
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2
boto3==1.28.57
